+++ romp/lib/dataset/image_base_relative.py
@@ -194,7 +194,7 @@ def test_image_relative_dataset(dataset,with_3d=False,with_smpl=False):
                 cv2.imwrite('{}/{}_{}_projection.jpg'.format(save_dir,_,img_bsname), image_kp2d_projection)
 
             for pinds, (person_center, subject_id) in enumerate(zip(person_centers,subject_ids)):
-                y,x = person_center.astype(np.int)
+                y,x = person_center.astype(np.int32)
                 if y>0 and x>0:
                     cv2.circle(image_kp2d, (x,y), 6, [0,0,255],-1)
                     text = '{}'.format(subject_id)
@@ -218,7 +218,7 @@ def test_image_relative_dataset(dataset,with_3d=False,with_smpl=False):
             #             centermap_3d = r['centermap_3d'][rind]
             #             visualize_3d_hmap(centermap_3d, '{}/{}_{}'.format(save_dir, _, rind))
 
-            person_centers_onmap = ((r['person_centers'][inds].numpy() + 1)/ 2.0 * (args().centermap_size-1)).astype(np.int)
+            person_centers_onmap = ((r['person_centers'][inds].numpy() + 1)/ 2.0 * (args().centermap_size-1)).astype(np.int32)
             positive_position = torch.stack(torch.where(r['centermap'][inds,0]==1)).permute(1,0)
 
         if with_smpl and r['valid_masks'][0,0,4]:
@@ -257,7 +257,7 @@ def test_image_relative_dataset(dataset,with_3d=False,with_smpl=False):
                 image = np.array(image).astype(np.uint8)
                 if len(person_centers) == len(trans):
                     for pinds, person_center in enumerate(person_centers):
-                        y,x = person_center.astype(np.int)
+                        y,x = person_center.astype(np.int32)
                         if y>0 and x>0:
                             cv2.circle(np.array(image), (x,y), 6, [0,0,255],-1)
                             text = '{:.2f}'.format(trans[pinds,2])
