+++ romp/lib/tracking/tracking_utils/visualization.py
@@ -70,7 +70,7 @@ def plot_detections(image, tlbrs, scores=None, color=(255, 0, 0), ids=None):
     text_scale = max(1, image.shape[1] / 800.)
     thickness = 2 if text_scale > 1.3 else 1
     for i, det in enumerate(tlbrs):
-        x1, y1, x2, y2 = np.asarray(det[:4], dtype=np.int)
+        x1, y1, x2, y2 = np.asarray(det[:4], dtype=np.int32)
         if len(det) >= 7:
             label = 'det' if det[5] > 0 else 'trk'
             if ids is not None:
